move serialize back and add io
